requirements.txt
pandas
groq
python_dotenv
pymongo[srv]
langchain
langchain_core
langchain_groq
langchain_openai
langchain_community
streamlit
google-generativeai